+++ switzerland/src/main/java/org/eqasim/switzerland/ovgk/OVGK.java
@@ -1,0 +1,5 @@
+package org.eqasim.switzerland.ovgk;
+
+public enum OVGK {
+	A, B, C, D, None
+}