+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-** (master) **
+- Add tools for ÖV Güteklasse to Switzerland scenario
 
 - Add Sé shape file